Added safety to the unique field in the Family model
The unique field in the Family model now has the unique attribute set too
make sure that a non-unique Family models isn't accidentally created.
A side effect of setting the unique attribute is implicitly settings the
db_index. Uniqueness checking also has a small resource cost because the
database has to spend time checking that the value is unique, and django
checks that the field is unique when doing model validation.
The editable attribute of the unique field is set to False to prevent
editing which would cause user confusion and possible other problems.