Added login functionality
Users are now required to login to view the family page and other family related pages.
User can be bound to a Person to allow them to login. This mean that there need to be created a user for each Person that should be able to login.
Added a rudementary login page.
ToDo: add a logout button on the family page, and a login button on the frontpage