fix(test-data): departments end date & address errors
Departments no longer uses a factory that creates a "closed_dtm". but is set to None. This fixes a bug that made departments closed.

Departments' address.region is now the region of their union.

Unions are now in the right region.

Also includes minor code additions to prepare for a refactoring of handle() and the UNIONS_TO_CREATE list.