refactor(test-data): "Region" no longer part of union names
Each union used to be named "Region ____". which made it easy to use the union's name as the name of the region too (since there's a union for each region).

However this made the names wierd, because the union for Syddanmark should be named "Syddanmark", not "Region Syddanmark". The commit refactors the UNIONS_TO_CREATE list to contain named tuples which has "union_name" and "region_name". This
fixes the wierd names and makes it possible to have a union be called something completely unrelated to the region its in, should you want that.