refactor(test-data): move waitlist setup to function
Setup of the waitlist, is moved into its own _setup_waitlist() function. It creates two children and puts them on a waitlist for two departments (which handle() supplies)

The function is very 'static' and only does what is required (no parameters for alternative options). Hopefully, moving the code to its own function. will provide an easier opportunity to expand it later on.

This update also fixes a bug with the help text.