fix(test-data): signup_closing is before script has run
The ActivityFactory made the signup_closing field be a date +/-100 days from when the script was run. This made some acitivities not show up in the activities overview of the website, since the signup had already closed. This is now fixed by setting signup_closed to a date 1-10 days in the future.